add entity graphs

+++ src/main/java/com/goapi/goapi/repo/finances/payment/UserBillPaymentRepository.java
@@ -1,7 +1,7 @@
 package com.goapi.goapi.repo.finances.payment;
 
-import com.goapi.goapi.domain.model.payment.UserBillPayment;
+import com.goapi.goapi.domain.model.finances.payment.UserBillPayment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserBillPaymentRepository extends JpaRepository<UserBillPayment, Long> {
+public interface UserBillPaymentRepository extends JpaRepository<UserBillPayment, Integer> {
 }
